[medium][extracted PodCallResult interface]
