+++ exe/src/main/kotlin/io/wavebeans/execution/config/Config.kt
@@ -1,0 +1,13 @@
+package io.wavebeans.execution.config
+
+import io.wavebeans.execution.medium.PodCallResultBuilder
+import io.wavebeans.execution.medium.PodCallResultWithSerializationBuilder
+
+object Config {
+
+    /**
+     * What builder to use build [io.wavebeans.execution.medium.PodCallResult]
+     */
+    fun podCallResultBuilder(): PodCallResultBuilder = PodCallResultWithSerializationBuilder()
+
+}
