+++ exe/src/main/kotlin/io/wavebeans/execution/podproxy/MergingPodProxy.kt
@@ -1,14 +1,11 @@
 package io.wavebeans.execution.podproxy
 
 import io.wavebeans.execution.*
-import io.wavebeans.execution.medium.PodCallResult
 import io.wavebeans.execution.pod.DEFAULT_PARTITION_SIZE
 import io.wavebeans.execution.pod.PodKey
 import io.wavebeans.lib.AnyBean
-import io.wavebeans.lib.Bean
 import io.wavebeans.lib.BeanParams
 import io.wavebeans.lib.BeanStream
-import java.util.concurrent.TimeUnit
 
 abstract class MergingPodProxy(
         override val forPartition: Int,
