+++ exe/src/test/kotlin/io/wavebeans/execution/podproxy/MergingPodProxySpec.kt
@@ -6,8 +6,6 @@ import assertk.assertions.isNotNull
 import com.nhaarman.mockitokotlin2.mock
 import io.wavebeans.execution.*
 import io.wavebeans.execution.medium.PodCallResult
-import io.wavebeans.execution.medium.SampleArray
-import io.wavebeans.execution.medium.nullableSampleArrayList
 import io.wavebeans.execution.pod.Pod
 import io.wavebeans.execution.pod.PodKey
 import io.wavebeans.lib.Sample
